Move developer tools into separate module
