+++ examples/bitcoin.rs
@@ -5,8 +5,8 @@ extern crate hex_literal;
 extern crate uint;
 
 use halo::{
-    is_satisfied, sha256::sha256, unpack_fe, AllocatedBit, AllocatedNum, Basic, Boolean, Circuit,
-    Coeff, ConstraintSystem, Field, Fp, LinearCombination, SynthesisError, UInt64,
+    dev::is_satisfied, sha256::sha256, unpack_fe, AllocatedBit, AllocatedNum, Basic, Boolean,
+    Circuit, Coeff, ConstraintSystem, Field, Fp, LinearCombination, SynthesisError, UInt64,
 };
 use sha2::{Digest, Sha256};
 use std::iter;
