+++ src/gadgets/uint64.rs
@@ -226,7 +226,7 @@ impl UInt64 {
 mod test {
     use super::UInt64;
     use crate::{
-        circuits::Circuit, fields::Fp, gadgets::boolean::Boolean, is_satisfied, Basic,
+        circuits::Circuit, dev::is_satisfied, fields::Fp, gadgets::boolean::Boolean, Basic,
         ConstraintSystem, SynthesisError,
     };
     use rand_core::{RngCore, SeedableRng};
